peplus.c: Minor update - NW
Set the French 32K IP0xxx poker games to the nonplus init. Explain in the header of the source file under Initialization : Method 3 about the secondary door that needs to be closed to play. Also made Joker Poker sets clones of the lowest numbered Joker Poker and did the same for the Standard Draw sets.